+++ R/utils.R
@@ -76,4 +76,21 @@ show_json <- function(x) {
 print.bq_bytes <- function(x, ...) {
   cat_line(prettyunits::pretty_bytes(unclass(x)))
 }
+
+#' Loosely adapted from https://github.com/googleapis/python-api-core/blob/master/google/api_core/client_info.py
+#' @noMd
+bqs_ua <- function() {
+  paste0(
+    "bigrquery",
+    utils::packageVersion("bigrquery"),
+    "r/",
+    R.version$major,
+    ".",
+    R.version$minor,
+    R.version$platform,
+    " grpc/",
+    paste0(grpc_version(), collapse = "_")
+  )
+}
+
 # nocov end